+++ src/pages/SupplierRequisitionPage.js
@@ -28,7 +28,7 @@ import {
   selectCurrentIndicatorCode,
   selectIndicatorTableColumns,
   selectIndicatorTableRows,
-} from './dataTableUtilities/selectors/indicatorSelectors';
+} from '../selectors/indicatorSelectors';
 import { getItemLayout, getPageDispatchers, PageActions } from './dataTableUtilities';
 import { ROUTES } from '../navigation/constants';
