+++ src/pages/dataTableUtilities/actions/pageActions.js
@@ -6,7 +6,7 @@
 import { UIDatabase } from '../../../database/index';
 import { MODAL_KEYS } from '../../../utilities';
 import { ACTIONS } from './constants';
-import { pageObjectSelector, pageStateSelector } from '../selectors/pageSelectors';
+import { pageObjectSelector, pageStateSelector } from '../../../selectors/pageSelectors';
 
 /**
  * Refreshes the underlying data array by slicing backingData.
