+++ src/pages/dataTableUtilities/actions/tableActions.js
@@ -10,7 +10,7 @@ import Settings from '../../../settings/MobileAppSettings';
 import { createRecord } from '../../../database/utilities/index';
 import { closeModal } from './pageActions';
 import { ACTIONS } from './constants';
-import { pageObjectSelector } from '../selectors/pageSelectors';
+import { pageObjectSelector } from '../../../selectors/pageSelectors';
 
 /**
  * Sorts the underlying data array by the key provided. Determines
